Solved irrigation two days in a row
Because of the order in which the different subroutines run, the
irrigation algorithm was running before potential root available water
(TRWUP) was updated and as a result the crops were irrigated in two
consecutive days.
Problem solved by explicitly telling the program not to irrigate two
days in a row.