Limited Irrigation works with hard coded data
New approach to limited irrigation. Different lines are entered in the X
file for each growth stage. In each line the user specifies all of the
automatic irrigation parameters and the growth stages where they should
be applied. User must enter growth stages as integers, and the GS should
be in the order they take place during plants life cycle.