Resolve issues discussed with Cheryl and Dr. Hoogenboom
Replace code using the go to function in IPSIM.for with DO WHILE
loop.
Change name of GINDEX variable to GSIRRIG.
Solve error in IRRIG.for line 135, match loop iterations with array
size.
Irrigation limit of blank, -99, and GS000 should be treated as no
limitation and zero should be treated as zero.
Resolve incompatibility with some old files (was reading 0 when AVWAT
was not specified)