Fixed issue related to generating output as fort.105
ISSUE:When in OUTPUT.CDE file were added additional *.csv files as
outputs after running CSM for all options for FMTOPT in file-X it was
generated fort.105 file with part for INFO.OUT file.
FIX: in info.for addded section ELSE when FIRST is FALSE. When missing
this case LUN was not associated with existing INFO.OUT file.