Add files via upload
This project is part of a doctoral research on pineapple cultivation, cultivar MD-2 in Costa Rica.

The files contained therein are those that to this date have been modified from the original Aloha_Pineapple model files, and others contained in the InputModule and Genotype folders, with the following purposes:
1. To add more phenological stages to the existing ones in the Aloha_Pineapple Model available in version 4.7.5.0 of DSSAT.
2. To add more cultivar coefficients.
3. To calibrate the model with data from 3 contrasting natural conditions in terms of temperature and solar radiation.

Additionally, the executable file compiled in Microsoft Visual Studio Community 2019 Version 16.7.1 has been added.

To observe the differences between this compilation and the one currently available in DSSAT, one must run one of the experiments available for pineapple and for example open the Overview.out file. Subsequently replace the PIALO047.CUL file in the Genotype folder with the one in this repository and then replace the dscsm047.exe file that is in the DSSAT folder with the one in this compilation, run the same experiment and compare the differences in the new Overview.out. Note that more stages have been added, which are vegetative and reproductive phenological stages.
Currently the work has focused on creating a structure in the Aloha_Pineapple model compatible with the additional data that will be collected in the research. The author is not an expert in programming and all this work has been possible thanks to the instructions shared by the DSSAT foundation. Any help and advice will be very welcome.